// src/_services/api/model/demandDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Gestão de Devolução e Produtividade
 * Documentação da API responsável por: gestão de produtividade, montagem e impressão de mapas de separação, sistema de devolução de estoque e contagem de inventário.
 * OpenAPI spec version: 1.0
 */
import type { DemandDtoIdTransportadora } from './demandDtoIdTransportadora';
import type { DemandDtoTelefone } from './demandDtoTelefone';
import type { DemandDtoQuantidadePaletes } from './demandDtoQuantidadePaletes';
import type { DemandDtoDoca } from './demandDtoDoca';
import type { DemandDtoConferenteId } from './demandDtoConferenteId';
import type { DemandDtoViagemId } from './demandDtoViagemId';
import type { DemandDtoStatus } from './demandDtoStatus';

export interface DemandDto {
  /** @minLength 1 */
  placa: string;
  /** @minLength 1 */
  motorista: string;
  idTransportadora?: DemandDtoIdTransportadora;
  telefone?: DemandDtoTelefone;
  cargaSegregada?: boolean;
  quantidadePaletes?: DemandDtoQuantidadePaletes;
  doca?: DemandDtoDoca;
  /** @minLength 1 */
  centerId: string;
  /** @minLength 1 */
  adicionadoPorId: string;
  conferenteId?: DemandDtoConferenteId;
  /** @minLength 1 */
  senha: string;
  viagemId?: DemandDtoViagemId;
  status: DemandDtoStatus;
  criadoEm: string;
  id: number;
}